Add `writeWritablePackets` and its test, and make `defaultMaxBufferSize` an int64

+++ engine.io/constants.go
@@ -9,8 +9,8 @@ import (
 const (
 	ProtocolVersion = parser.ProtocolVersion
 
-	defaultMaxBufferSize  int = 1e6 // 100 MB
-	defaultPingTimeout        = time.Second * 20
-	defaultPingInterval       = time.Second * 25
-	defaultUpgradeTimeout     = time.Second * 10
+	defaultMaxBufferSize  int64 = 1e6 // 100 MB
+	defaultPingTimeout          = time.Second * 20
+	defaultPingInterval         = time.Second * 25
+	defaultUpgradeTimeout       = time.Second * 10
 )
